requirements.txt
# Movie Recommendation System - Dependencies
# Python 3.9+

# Core
numpy>=1.24,<2
pandas>=2.0
pyyaml>=6.0
scikit-learn>=1.3

# Implicit collaborative filtering
implicit>=0.7

# Data storage
pyarrow>=14.0  # for parquet

# API
fastapi>=0.104
uvicorn[standard]>=0.24

# Dev & testing
pytest>=7.4
pytest-cov>=4.1
httpx>=0.25  # for API tests
black>=23.0
ruff>=0.1
